requirements.txt
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
django-bootstrap3==21.2
PyMySQL==1.0.2
requests==2.27.1
loguru